requirements.txt
requests>=2.31.0
feedparser>=6.0.10
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0